DataDump full refactor
