+++ src/main/java/org/recap/model/export/FullDataDumpCallable.java
@@ -2,7 +2,6 @@ package org.recap.model.export;
 
 import org.recap.model.jpa.BibliographicEntity;
 import org.recap.repository.BibliographicDetailsRepository;
-import org.recap.util.datadump.BibRecordConverter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Scope;
