+++ src/main/java/org/recap/report/FTPDataDumpSuccessReportGenerator.java
@@ -4,11 +4,8 @@ import org.apache.camel.ProducerTemplate;
 import org.apache.commons.io.FilenameUtils;
 import org.recap.ReCAPConstants;
 import org.recap.model.csv.DataDumpSuccessReport;
-import org.recap.model.csv.ReCAPCSVSuccessRecord;
-import org.recap.model.csv.SuccessReportReCAPCSVRecord;
 import org.recap.model.jpa.ReportEntity;
-import org.recap.util.DataDumpSuccessReportGenerator;
-import org.recap.util.ReCAPCSVSuccessRecordGenerator;
+import org.recap.util.datadump.DataDumpSuccessReportGenerator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.util.CollectionUtils;
